CLEVR_192_neurips/requirements.txt
torchvision==0.6.0+cu101
numpy==1.16.4
torch==1.5.0+cu101
matplotlib==3.1.0
six==1.12.0
data==0.4
Pillow==7.1.2
